# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the messaging service",
        "title": "Messaging API",
        "contact": {
            "name": "Çağla Çolak",
            "email": "caglaccolak@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/sent-messages": {
            "get": {
                "description": "Returns a list of all messages that have been sent.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messaging"
                ],
                "summary": "Retrieve sent messages",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/start-messages": {
            "post": {
                "description": "Starts the scheduler that sends unsent messages every 2 minutes.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scheduler"
                ],
                "summary": "Start the scheduler",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/stop-messages": {
            "post": {
                "description": "Stops the currently running scheduler for sending messages.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scheduler"
                ],
                "summary": "Stop the scheduler",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Hello, world!"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "recipient": {
                    "type": "string",
                    "example": "+905383311137"
                },
                "sent_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "sent"
                }
            }
        }
    }
}